emsdk 5.0.2

+++ src/GeraNESApp/EmscriptenBindings.cpp
@@ -3,18 +3,19 @@
 #ifdef __EMSCRIPTEN__
 
 #include <emscripten.h>
+#include <cstdint>
 
 extern "C" {
 
-    void EMSCRIPTEN_KEEPALIVE processUploadedFile(int handler, const char* fileName, size_t fileSize, const uint8_t* fileContent) {
+    void EMSCRIPTEN_KEEPALIVE processUploadedFile(intptr_t handler, const char* fileName, size_t fileSize, const uint8_t* fileContent) {
         reinterpret_cast<GeraNESApp*>(handler)->processUploadedFile(fileName, fileSize, fileContent);
     }
 
-    void EMSCRIPTEN_KEEPALIVE restartAudioModule(int handler) {
+    void EMSCRIPTEN_KEEPALIVE restartAudioModule(intptr_t handler) {
         reinterpret_cast<GeraNESApp*>(handler)->restartAudioModule();
     }
 
-    void EMSCRIPTEN_KEEPALIVE onSessionImportComplete(int handler) {
+    void EMSCRIPTEN_KEEPALIVE onSessionImportComplete(intptr_t handler) {
         reinterpret_cast<GeraNESApp*>(handler)->onSessionImportComplete();
     }
